consume my own module
